merging

+++ application/models/Maba_dashboard_model.php
@@ -9,9 +9,9 @@ class Maba_dashboard_model extends CI_Model {
    	$this->db=$db;
    }
 
-   private function angkatan($angk)
+   private function thnlls($angk)
    {
-    $curYear = date('Y')-3;
+    $curYear = date('Y');
 
      $ang=array();
      for ($i=2008; $i <= $curYear ; $i++) { 
@@ -37,15 +37,13 @@ class Maba_dashboard_model extends CI_Model {
 
    public function baca_data()
    {
-     $id_wisuda = $this->session->userdata('id_wisuda');
-     
+     $id_peserta = $this->session->userdata('id_peserta');
 
-
-     $tmp=$this->db['wisudawan']->getdata('id_wisuda="'.$id_wisuda.'"');
+     $tmp=$this->db['maba']->getdata('id_peserta="'.$id_peserta.'"');
      
-     $arr_ang=$this->angkatan($tmp[0]['angkatan']);     
-     $data['drop_ang']=$this->build_dropdown($arr_ang,array(0,1),'','--- Pilih Angkatan ---');    
-     $data['drop_ang']= "<option value='".$tmp[0]['angkatan']."' selected='selected' >".$tmp[0]['angkatan']."</option>".$data['drop_ang'];
+     $arr_thnlls=$this->thnlls($tmp[0]['thnlls']);     
+     $data['drop_thnlls']=$this->build_dropdown($arr_thnlls,array(0,1),'','--- Pilih Tahun Lulus ---');    
+     $data['drop_thnlls']= "<option value='".$tmp[0]['thnlls']."' selected='selected' >".$tmp[0]['thnlls']."</option>".$data['drop_thnlls'];
 
      $tmp_prodi=$this->db['prodi']->getdata('id_prodi="'.$tmp[0]['id_prodi'].'"');     
      $tmp_fak=$this->db['fakultas']->getdata('id_fak="'.$tmp_prodi[0]['fak_prodi'].'"');
@@ -58,24 +56,31 @@ class Maba_dashboard_model extends CI_Model {
      $data['drop_prodi']=$this->build_dropdown($arr_prodi,array('id_prodi','nm_prodi'),'Prodi. ','--- Pilih Prodi ---');
      $data['drop_prodi']= "<option value='".$tmp_prodi[0]['id_prodi']."' selected='selected' >Prodi. ".$tmp_prodi[0]['nm_prodi']."</option>".$data['drop_prodi'];
 
+     $data['id_peserta']=$tmp[0]['id_peserta'];
      $data['ktp']=$tmp[0]['ktp'];
-     $data['nama']=$tmp[0]['nama'];
+     $data['nama']=$tmp[0]['nm'];
      $data['jk']=$tmp[0]['jk'];
-     $data['tmpt_lahir']=$tmp[0]['tmpt_lahir'];
-     $data['tgl_lahir']=date("d-m-Y", strtotime($tmp[0]['tgl_lahir']));
+     $data['tmpt_lahir']=$tmp[0]['tmplhr'];
+     $data['tgl_lahir']=date("d-m-Y", strtotime($tmp[0]['tgllhr']));
      $data['alamat']=empty($tmp[0]['alamat']) ? '' : $tmp[0]['alamat'];
-     $data['hp']=$tmp[0]['hp'];
+     $data['asal']=empty($tmp[0]['asalsma']) ? '' : $tmp[0]['asalsma'];
+     $data['email']=empty($tmp[0]['email']) ? '' : $tmp[0]['email'];
+     $data['hp']=$tmp[0]['hp'];     
      $data['photo']=$tmp[0]['photo'];
-     $data['tgl_byr']= is_null($tmp[0]['tgl_byr'])  ? '' : date("d-m-Y", strtotime($tmp[0]['tgl_byr']));
+     $data['tgl_trans']= is_null($tmp[0]['tgltrans'])  ? '' : date("d-m-Y", strtotime($tmp[0]['tgltrans']));
      $data['kwitansi']=$tmp[0]['kwitansi'];
+     $data['bank']=empty($tmp[0]['nm_bank']) ? '' : $tmp[0]['nm_bank'];
+     $data['kelas']=empty($tmp[0]['kelas']) ? '' : $tmp[0]['kelas'];     
+    
+     $data['ver']=$tmp[0]['verified'];
 
-     $data['nim']=$tmp[0]['nim'];
-     
-     //$data['ipk']=$tmp[0]['ipk'];
-     //$data['tgl_lls']= is_null($tmp[0]['tgl_lls'])  ? '' : date("d-m-Y", strtotime($tmp[0]['tgl_lls']));
+     $tmp = $this->db['priode']->getrek();
 
-     $data['jdl_skripsi']=$tmp[0]['jdl_skripsi'];
-     $data['ver']=$tmp[0]['ver'];
+     $data['konf_msg']='Biaya pendaftaran Rp. '.number_format($tmp['byr'],2,',','.').'<br> ditransfer ke <br>'.
+                       'No Rekening : '.$tmp['rek'].'<br>'.
+                       'Bank        : '.$tmp['bank'].'<br>'.
+                       'Atas Nama   : '.$tmp['an'];
+
 
      return $data;
 
@@ -123,7 +128,7 @@ class Maba_dashboard_model extends CI_Model {
     }else{
 
       if($this->db['priode']->islogin()==0){
-        $data['msg']="<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Perubahan Data Setelah Wisuda Tidak Diijinkan !!!</p> </div>";
+        $data['msg']="<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Perubahan Data Setelah Priode Pendaftaran Tidak Diijinkan !!!</p> </div>";
       }else{
        $data['isbuka']=1;
       }
@@ -139,14 +144,14 @@ class Maba_dashboard_model extends CI_Model {
    	 if($login=='login')
    	 {
            
-        $tmp=$this->db['wisudawan']->getdata('user_name="'.$un.'" and user_pass=md5("'.$psw.'")'); 
+        $tmp=$this->db['maba']->getdata('user="'.$un.'" and pass=md5("'.$psw.'")'); 
         
         if(empty($tmp))
         {
           $data['msg']="<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Username atau Password Keliru !!!</p> </div>";
         }else{
         	$newdata = array(
-               'id_wisuda'  => $tmp[0]['id_wisuda'],
+               'id_peserta'  => $tmp[0]['id_peserta'],
                'lg_time' => date('Y-m-d H:i:s'),
                'logged_in' => TRUE
              );
@@ -166,19 +171,19 @@ class Maba_dashboard_model extends CI_Model {
      $date = date('Y-m-d');
      $data['isbuka']= $date >= $tmp[0]['awal'] && $date <= $tmp[0]['akhir'];
 
-     $tmp=$this->db['wisudawan']->getdata("ktp='$data[ktp]' and tgl_lahir='$data[tgl_lahir]' and jk='$data[jk]'");
+     $tmp=$this->db['maba']->getdata("ktp='$data[ktp]' and tgllhr='$data[tgl_lahir]' and jk='$data[jk]'");
      $msg='';
      if(!empty($tmp)){
-        $tmp1['id_wisuda']=$tmp[0]['id_wisuda'];
-        $tmp=$this->db['wisudawan']->getdata("user_name='$data[user_name]' and ktp<>'$data[ktp]'");
+        $tmp1['id_peserta']=$tmp[0]['id_peserta'];
+        $tmp=$this->db['maba']->getdata("user='$data[user_name]' and ktp<>'$data[ktp]'");
         if(!empty($tmp)){
              $data['kd']=0;
              $msg = "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Username ada dalam database !!!</p> </div>";
              }
              else{
-               $tmp1['user_name']=$data['user_name'];
-               $tmp1['user_pass']=md5($data['user_pass']);
-               $this->db['wisudawan']->updatedata($tmp1);
+               $tmp1['user']=$data['user_name'];
+               $tmp1['pass']=md5($data['user_pass']);
+               $this->db['maba']->updatedata($tmp1);
                $msg="<div class='callout callout-info'><h4>Pemberitahuan</h4><p>User name dan Password berhasil diganti, silahkan login !!!</p> </div>";
                $data['kd']=1;
              }
@@ -190,58 +195,37 @@ class Maba_dashboard_model extends CI_Model {
      return $data; 
    }
 
-   public function logout($id_wisuda,$lg_time)
+   public function logout($id_peserta,$lg_time)
    {
       $out_time = date('Y-m-d H:i:s');
-      $tmp=$this->db['log']->updatedata(array('id_wisuda'=>$id_wisuda,'lg_time'=>$lg_time,'out_time'=>$out_time));
+      $tmp=$this->db['log']->updatedata(array('id_peserta'=>$id_peserta,'lg_time'=>$lg_time,'out_time'=>$out_time));
    }
 
-   public function updatedatawisudawan($data)
+   public function updatedatamaba($data)
    {
-     $tmp=$this->db['wisudawan']->getdata("ktp='$data[ktp]' and id_wisuda<>'$data[id_wisuda]'");
+     $tmp=$this->db['maba']->getdata("ktp='$data[ktp]' and id_peserta<>'$data[id_peserta]'");
          if(!empty($tmp)){
-             return "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Calon wisudawan dengan ktp/nik = $data[ktp], sudah ada !!!</p> </div>"; 
+             return "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan ktp/nik = $data[ktp], sudah ada !!!</p> </div>"; 
          }else{
 
-             $tmp=$this->db['wisudawan']->getdata("nim='$data[nim]' and id_wisuda<>'$data[id_wisuda]'");
-                if(!empty($tmp)){
-                      return "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Calon wisudawan dengan nim = $data[nim], sudah ada !!!</p> </div>"; 
-                }else{
-
-                    $is_simpan=1;
-                    if(isset($data['user_name'])){ 
-                       $tmp=$this->db['wisudawan']->getdata("user_name='$data[user_name]' and id_wisuda<>'$data[id_wisuda]'");
-                       if(!empty($tmp)){
-                         $is_simpan=0;
-                         return "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Wisudawan dengan user_name = $data[user_name], sudah ada !!!</p> </div>"; 
-                       }
-                    }   
-
+             
+                   
                        if(isset($data['photo'])){
                             if(file_exists('./assets/photo/'.basename($data['photo']))){
                                  $ext = explode('.',basename($data['photo']));
-                                 rename('./assets/photo/'.basename($data['photo']),'./assets/photo/photo_'.$data['id_wisuda'].'.'.$ext[1]);
-                                 $data['photo']=base_url().'assets/photo/photo_'.$data['id_wisuda'].'.'.$ext[1];
+                                 rename('./assets/photo/'.basename($data['photo']),'./assets/photo/'.$data['id_peserta'].'.'.$ext[1]);
+                                 $data['photo']=$data['id_peserta'].'.'.$ext[1];
                              }else{
                                 unset($data['photo']);
                              }
                            }
 
-                           if(isset($data['kwitansi'])){
-                              if(file_exists('./assets/photo/'.basename($data['kwitansi']))){
-                                 $ext = explode('.',basename($data['kwitansi']));
-                                 rename('./assets/photo/'.basename($data['kwitansi']),'./assets/photo/kwitansi_'.$data['id_wisuda'].'.'.$ext[1]);
-                                 $data['kwitansi']=base_url().'assets/photo/kwitansi_'.$data['id_wisuda'].'.'.$ext[1];
-                             }else{
-                              unset($data['kwitansi']);
-                             }
-                           }
+                   
+                      
+                         $this->db['maba']->updatedata($data);
 
-                      if($is_simpan==1){ 
-                         $this->db['wisudawan']->updatedata($data);
-
-                         $id_wisuda = $this->session->userdata('id_wisuda');
-                         $nm_file = 'temp_'.$id_wisuda;
+                         $id_peserta = $this->session->userdata('id_peserta');
+                         $nm_file = 'temp_'.$id_peserta;
 
                          $file = directory_map('./assets/photo/');
 
@@ -250,18 +234,91 @@ class Maba_dashboard_model extends CI_Model {
                             foreach ($file as $nmfile) {
                                $ext = explode('.',basename($nmfile));
                                $tmp = explode('_',$ext[0]);
-                               if(($tmp[0].'_'.$tmp[1])==$nm_file)
-                               {
-                                 unlink('./assets/photo/'.$nmfile);   
+                               
+                               if(isset($tmp[1])){
+                                 if(($tmp[0].'_'.$tmp[1])==$nm_file)
+                                 {
+                                    unlink('./assets/photo/'.$nmfile);   
+                                 }
+                               }  
+
+                             } 
+                         }
+
+
+                         return "<div class='callout callout-info'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan id peserta = $data[id_peserta], berhasil di update !!!</p> </div>"; 
+                         
+                 
+         }
+   }
+
+   public function updatedatapil($data)
+   {
+     
+     $this->db['maba']->updatedata($data);
+
+     return "<div class='callout callout-info'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan id peserta = $data[id_peserta], berhasil di update !!!</p> </div>"; 
+                       
+        
+   }
+
+   public function konf($data)
+   {
+     
+
+                           if(isset($data['kwitansi'])){
+                              if(file_exists('./assets/photo/'.basename($data['kwitansi']))){
+                                 $ext = explode('.',basename($data['kwitansi']));
+                                 rename('./assets/photo/'.basename($data['kwitansi']),'./assets/photo/kwitansi_'.$data['id_peserta'].'.'.$ext[1]);
+                                 $data['kwitansi']='kwitansi_'.$data['id_peserta'].'.'.$ext[1];
+                             }else{
+                              unset($data['kwitansi']);
+                             }
+                           }
+                         
+                         $data['konf']=1;
+                         $data['tglkonf']=date('Y-m-d'); 
+                     
+                         $this->db['maba']->updatedata($data);
+
+                         $id_peserta = $this->session->userdata('id_peserta');
+                         $nm_file = 'temp_'.$id_peserta;
+
+                         $file = directory_map('./assets/photo/');
+
+                         if(!empty($file))                          
+                         {
+                            foreach ($file as $nmfile) {
+                               $ext = explode('.',basename($nmfile));
+                               $tmp = explode('_',$ext[0]);
+                               if(isset($tmp[1])){
+                                 if(($tmp[0].'_'.$tmp[1])==$nm_file)
+                                 {
+                                    unlink('./assets/photo/'.$nmfile);   
+                                 }
                                }
                              } 
                          }
 
 
-                         return "<div class='callout callout-info'><h4>Pemberitahuan</h4><p>Calon Wisudawan dengan id wisuda = $data[id_wisuda], berhasil di update !!!</p> </div>"; 
-                      }   
-               }    
-         }
+                         return "<div class='callout callout-info'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan id peserta = $data[id_peserta], berhasil di update !!!</p> </div>"; 
+ 
+   }
+
+
+   public function rubahuserpass($data)
+   {
+                    if(isset($data['user'])){ 
+                       $tmp=$this->db['maba']->getdata("user='$data[user]' and id_peserta<>'$data[id_peserta]'");
+                       if(!empty($tmp)){                         
+                         return "<div class='callout callout-danger'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan user_name = $data[user], sudah ada !!!</p> </div>"; 
+                       }else{
+                          $this->db['maba']->updatedata($data);
+                          return "<div class='callout callout-info'><h4>Pemberitahuan</h4><p>Mahasiswa baru dengan id peserta = $data[id_peserta], berhasil di update !!!</p> </div>"; 
+
+                       }
+                    }               
+                      
    }
